+++ tensordict/_td.py
@@ -2666,55 +2666,6 @@ class TensorDict(TensorDictBase):
 
         return memmap_tensor
 
-    def to(self, *args, **kwargs: Any) -> T:
-        non_blocking = kwargs.pop("non_blocking", None)
-        pin_memory = kwargs.pop("pin_memory", False)
-        num_threads = kwargs.pop("num_threads", None)
-
-        device, dtype, _, convert_to_format, batch_size = _parse_to(*args, **kwargs)
-        result = self
-
-        if device is not None and dtype is None and device == self.device:
-            return result
-
-        if non_blocking is None:
-            sub_non_blocking = True
-            non_blocking = False
-        else:
-            sub_non_blocking = non_blocking
-
-        if convert_to_format is not None:
-
-            def to(tensor):
-                return tensor.to(
-                    device,
-                    dtype,
-                    non_blocking=sub_non_blocking,
-                    convert_to_format=convert_to_format,
-                )
-
-        else:
-
-            def to(tensor):
-                return tensor.to(
-                    device=device, dtype=dtype, non_blocking=sub_non_blocking
-                )
-
-
-        apply_kwargs = {}
-        if device is not None or dtype is not None:
-            if pin_memory and num_threads != 0:
-                result = self._multithread_apply_nest(lambda x: x.pin_memory(), num_threads=num_threads, call_when_done=to)
-            else:
-                apply_kwargs["device"] = device if device is not None else self.device
-                apply_kwargs["batch_size"] = batch_size
-                result = result._fast_apply(to, propagate_lock=True, **apply_kwargs)
-        if batch_size is not None:
-            result.batch_size = batch_size
-        if device is not None and sub_non_blocking and not non_blocking:
-            self._sync_all()
-        return result
-
     def where(self, condition, other, *, out=None, pad=None):
         if _is_tensor_collection(other.__class__):
 
@@ -3315,9 +3266,15 @@ class _SubTensorDict(TensorDictBase):
         return self
 
     def to(self, *args, **kwargs: Any) -> T:
-        device, dtype, non_blocking, convert_to_format, batch_size = _parse_to(
-            *args, **kwargs
-        )
+        (
+            device,
+            dtype,
+            non_blocking,
+            convert_to_format,
+            batch_size,
+            pin_memory,
+            num_threads,
+        ) = _parse_to(*args, **kwargs)
         result = self
 
         if device is not None and dtype is None and device == self.device:
