+++ tensordict/nn/params.py
@@ -12,7 +12,7 @@ import weakref
 from concurrent.futures import Future, ThreadPoolExecutor
 from copy import copy
 from functools import wraps
-from typing import Any, Callable, Iterator, List, OrderedDict, Sequence, Type
+from typing import Any, Callable, Dict, Iterator, List, OrderedDict, Sequence, Type
 
 import torch
